requirements.txt
# Core web framework
fastapi==0.135.1
uvicorn[standard]==0.34.0      # ASGI server with standard extras

# Type hints & validation
pydantic==2.12.5

# Async HTTP client (for future Grok/Oura calls)
httpx==0.28.1

# Environment variables
python-dotenv==1.2.2

# JWT for session after Oura OAuth (Phase 1 multi-user)
PyJWT==2.10.1

# Observability & rate limiting
slowapi==0.1.9
sentry-sdk==2.54.0

# Testing
pytest==9.0.2

